Move gesture handling out of cursor.c
