+++ src/input/meson.build
@@ -1,6 +1,7 @@
 labwc_sources += files(
+  'cursor.c',
+  'gestures.c',
   'keyboard.c',
   'key-state.c',
   'touch.c',
-  'cursor.c',
 )
